Collect tests for stats.serlo.org in own file
