+++ __tests__/__utils__/services/index.ts
@@ -25,5 +25,4 @@ export * from './database'
 export * from './frontend'
 export * from './packages'
 export * from './serlo'
-export * from './stats'
 export * from './utils'
